requirements.txt
# Core dependencies
pandas==2.2.3
numpy>=1.23.0
seaborn>=0.12.0
torch>=2.0.0
scikit-learn>=1.0.2
pot>=0.9.0

# Additional ML libraries for extended model support
xgboost>=1.7.0
lightgbm>=3.3.0

# Visualization and analysis
matplotlib>=3.5.0
plotly>=5.0.0

# Optimization libraries
scipy>=1.9.0
joblib>=1.2.0
scikit-optimize>=0.9.0

# Image processing and additional tools
Pillow>=8.3.0
tqdm>=4.62.0
mlxtend>=0.19.0

# Development dependencies (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
pytest>=7.0.0
